Improve upgrade handling.
Upgrade jobs were previously being tagged in a dubious way. They would
sometimes get stuck in a loop where new upgrade jobs would be started
after the current job finished.

Now they are tracked by the UpgradeHandler differently. The handler
makes an effort to find the latest upgrade and make sure it's actually
still running before setting it as its _resource member.

Furthermore, the previous version of the UpgradeHandler didn't verify
that the job it found actually belonged to the OdooInstance it was part
of. We now filter jobs by UID when loading them.