Remove IngressRouteHttp - no longer needed with global redirection.
The redirection via middleware "seemed" to be working but when
specifying http://www.bemade.org (or any other site), the connection
failed with a 404 returned. After much investigating, I found no way to
get it working other than to turn on global redirection from web to
websecure on the Traefik entrypoint itself. This is a change for the
better anyway, because now no raw HTTP traffic is allowed.